Fixed mixed domain tests, and use live domain.xml configuration for it.
Using the live configuration avoids the eternal problem of these tests
becoming stale. The DomainAdjustors trim the config down into something
more manageable, and make necessary changes to get it to work on the
targetted legacy version. This means that the huge domain.xml file can
now be removed, and only one host-slave.xml file is needed.

While here, I have changed the structure since I was never happy with
it, it made it hard to add tests for new versions. This new structure is
clearer to me anyway.